Trait loupe::MemoryUsageTracker

pub trait MemoryUsageTracker {
    fn track(&mut self, address: *const ()) -> bool;
}

Represent a bucket that can track memory addresses that have already been visited by MemoryUsage.

Required methods

fn track(&mut self, address: *const ()) -> bool

When first called on a given address returns true, false otherwise.